requirements.txt
click==8.1.3
Flask==2.2.2
greenlet==2.0.1
libsass==0.22.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
msgpack==1.0.4
peewee==3.15.4
PyYAML==6.0
Werkzeug==2.2.2
